memory/my_travel_agent/requirements.txt
google-adk
google-genai
google-cloud-aiplatform
python-dotenv
